refactor: Address missed comments
