--- client/src/components/ProfileSettings/EditProfile/useStyles.ts
@@ -1,11 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(() => ({
-  box: {
-    backgroundColor: 'white',
-    height: '100%',
-    width: '100%',
-  },
-}));
-
-export default useStyles;
