+++ client/src/pages/ProfileDetails/useStyles.ts
@@ -1,0 +1,16 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    backgroundColor: '#FAFAFA',
+  },
+  sitterProfileCard: {
+    minWidth: '60vw',
+    flex: 1,
+  },
+  bookSitterCard: {
+    flex: 1,
+  },
+}));
+
+export default useStyles;
